hotusa integration

+++ app/controllers/Owners.java
@@ -31,7 +31,7 @@ public class Owners extends Controller{
 		render(deal);
 	}
 	
-	public static void save(Long id, Integer quantity, Integer salePriceCents, boolean breakfastIncluded) {
+	public static void save(Long id, Integer quantity, Float salePriceCents, boolean breakfastIncluded) {
 		Logger.debug("breakfast: " + breakfastIncluded);
 	    Deal deal;
 	    // Retrieve post
